requirements.txt
colorama==0.4.6
paho-mqtt==1.6.1
pyftdi==0.54.0
pymodbus==3.3.2
pyserial==3.5
pyudev==0.24.0
pyusb==1.2.1
PyHamcrest==2.0.4
aiofiles==23.2.1
aiomonitor==0.6.0
aioserial==1.3.1
